move ineven indent rule and rename it to uneven rule
