+++ README.rst
@@ -139,7 +139,7 @@ Main features
     --reports rules_by_id,rules_by_error_type
 
     Issues by ids:
-    W0902 (ineven-indent)                : 5
+    W1007 (uneven-indent)                : 5
     E0904 (nested-for-loop)              : 4
     W0302 (not-capitalized-keyword-name) : 4
 
@@ -147,7 +147,7 @@ Main features
 
 - configurable lint rules::
 
-    --configure 0507:line_length:100 -c ineven-indent:severity:W
+    --configure 0507:line_length:100 -c uneven-indent:severity:W
 
 - external rules support::
